Address critical review: 64-bit VM, timeouts, TypeScript parity, safety
#11 Expression VM uses 64-bit stack (uint64_t) for all values including
cycles and hits counters. Prevents overflow after ~35 minutes at 2MHz.
Constants also stored as uint64_t.

#1 ListBreakpoints/ListWatchpoints only read hit_count when the machine
is paused, avoiding a data race with the emulation loop thread.

#2 TypeScript debugger client updated to full parity: Breakpoint and
WatchpointInfo interfaces include condition, stopCounterpart, hitCount.
addBreakpoint accepts endAddress/condition/stopCounterpart. addWatchpoint
accepts condition/stopCounterpart. toBreakpoint uses startAddress.

#3 ParasiteRunner gains add_watchpoint_entry/clear_watchpoint_entries
for C++ test code parity with Machine.

#4 Timeouts on all blocking client methods:
- Python: DEFAULT_TIMEOUT=30s on watch_execution_state, wait_for_stop,
  run_until. Raises DebuggerError on expiry via gRPC deadline.
- TypeScript: 30000ms default on waitForStop, runUntil. Raises
  DebuggerError on expiry via setTimeout.

#5 in_run_ flag set/cleared in step_instruction() on both Machine and
ParasiteRunner, ensuring wait_until_idle() is fully protective.

#6 Parasite reset coordination documented (client-side sequencing).

#7-9 Future work documented: server-side disassembly, stack traces,
memory access trace stream.

#10 Register state in stop events: decision recorded to defer (events
should stay lightweight; clients fetch registers on demand).

+++ src/core/include/beebium/debugger/Expression.hpp
@@ -59,7 +59,7 @@ enum class ExprOp : uint8_t {
 // Compiled expression bytecode.
 struct CompiledExpression {
     std::vector<ExprOp> ops;
-    std::vector<uint32_t> constants;  // indexed by position in ops stream
+    std::vector<uint64_t> constants;  // indexed by position in ops stream
 
     // Source text (for display/debugging)
     std::string source;
@@ -80,7 +80,7 @@ using PeekFunction = uint8_t(*)(void* context, uint16_t addr);
 
 // Evaluate a compiled expression against CPU state.
 // Returns the top-of-stack value (truthy if non-zero).
-uint32_t evaluate(const CompiledExpression& expr,
+uint64_t evaluate(const CompiledExpression& expr,
                   const ExprCpuState& cpu,
                   PeekFunction peek, void* peek_context);
